requirements.txt
apache-airflow==2.8.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
idna==3.6
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pathspec==0.12.1
pipreqs==0.4.13
platformdirs==4.1.0
py4j==0.10.9.7
pyspark==3.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
yarg==0.1.9
